Forward `tqdm` constructor arguments to `Progress` component (#4475)
* Forward tqdm constructor arguments to Progress component


* lint

* reorder args

* added tests

---------
