+++ test/test_blocks.py
@@ -1483,148 +1483,6 @@ class TestEvery:
         assert "At step 9" not in captured.out
 
 
-class TestProgressBar:
-    @pytest.mark.asyncio
-    async def test_progress_bar(self):
-        from tqdm import tqdm
-
-        with gr.Blocks() as demo:
-            name = gr.Textbox()
-            greeting = gr.Textbox()
-            button = gr.Button(value="Greet")
-
-            def greet(s, prog=gr.Progress()):
-                prog(0, desc="start")
-                time.sleep(0.25)
-                for _ in prog.tqdm(range(4), unit="iter"):
-                    time.sleep(0.25)
-                time.sleep(1)
-                for _ in tqdm(["a", "b", "c"], desc="alphabet"):
-                    time.sleep(0.25)
-                return f"Hello, {s}!"
-
-            button.click(greet, name, greeting)
-        demo.queue(max_size=1).launch(prevent_thread_lock=True)
-
-        async with websockets.connect(
-            f"{demo.local_url.replace('http', 'ws')}queue/join"
-        ) as ws:
-            completed = False
-            progress_updates = []
-            while not completed:
-                msg = json.loads(await ws.recv())
-                if msg["msg"] == "send_data":
-                    await ws.send(json.dumps({"data": [0], "fn_index": 0}))
-                if msg["msg"] == "send_hash":
-                    await ws.send(json.dumps({"fn_index": 0, "session_hash": "shdce"}))
-                if msg["msg"] == "progress":
-                    progress_updates.append(msg["progress_data"])
-                if msg["msg"] == "process_completed":
-                    completed = True
-                    break
-        print(progress_updates)
-        assert progress_updates == [
-            [
-                {
-                    "index": None,
-                    "length": None,
-                    "unit": "steps",
-                    "progress": 0.0,
-                    "desc": "start",
-                }
-            ],
-            [{"index": 0, "length": 4, "unit": "iter", "progress": None, "desc": None}],
-            [{"index": 1, "length": 4, "unit": "iter", "progress": None, "desc": None}],
-            [{"index": 2, "length": 4, "unit": "iter", "progress": None, "desc": None}],
-            [{"index": 3, "length": 4, "unit": "iter", "progress": None, "desc": None}],
-            [{"index": 4, "length": 4, "unit": "iter", "progress": None, "desc": None}],
-        ]
-
-    @pytest.mark.asyncio
-    async def test_progress_bar_track_tqdm(self):
-        from tqdm import tqdm
-
-        with gr.Blocks() as demo:
-            name = gr.Textbox()
-            greeting = gr.Textbox()
-            button = gr.Button(value="Greet")
-
-            def greet(s, prog=gr.Progress(track_tqdm=True)):
-                prog(0, desc="start")
-                time.sleep(0.25)
-                for _ in prog.tqdm(range(4), unit="iter"):
-                    time.sleep(0.25)
-                time.sleep(1)
-                for _ in tqdm(["a", "b", "c"], desc="alphabet"):
-                    time.sleep(0.25)
-                return f"Hello, {s}!"
-
-            button.click(greet, name, greeting)
-        demo.queue(max_size=1).launch(prevent_thread_lock=True)
-
-        async with websockets.connect(
-            f"{demo.local_url.replace('http', 'ws')}queue/join"
-        ) as ws:
-            completed = False
-            progress_updates = []
-            while not completed:
-                msg = json.loads(await ws.recv())
-                if msg["msg"] == "send_data":
-                    await ws.send(json.dumps({"data": [0], "fn_index": 0}))
-                if msg["msg"] == "send_hash":
-                    await ws.send(json.dumps({"fn_index": 0, "session_hash": "shdce"}))
-                if (
-                    msg["msg"] == "progress" and msg["progress_data"]
-                ):  # Ignore empty lists which sometimes appear on Windows
-                    progress_updates.append(msg["progress_data"])
-                if msg["msg"] == "process_completed":
-                    completed = True
-                    break
-        assert progress_updates == [
-            [
-                {
-                    "index": None,
-                    "length": None,
-                    "unit": "steps",
-                    "progress": 0.0,
-                    "desc": "start",
-                }
-            ],
-            [{"index": 0, "length": 4, "unit": "iter", "progress": None, "desc": None}],
-            [{"index": 1, "length": 4, "unit": "iter", "progress": None, "desc": None}],
-            [{"index": 2, "length": 4, "unit": "iter", "progress": None, "desc": None}],
-            [{"index": 3, "length": 4, "unit": "iter", "progress": None, "desc": None}],
-            [{"index": 4, "length": 4, "unit": "iter", "progress": None, "desc": None}],
-            [
-                {
-                    "index": 0,
-                    "length": 3,
-                    "unit": "steps",
-                    "progress": None,
-                    "desc": "alphabet",
-                }
-            ],
-            [
-                {
-                    "index": 1,
-                    "length": 3,
-                    "unit": "steps",
-                    "progress": None,
-                    "desc": "alphabet",
-                }
-            ],
-            [
-                {
-                    "index": 2,
-                    "length": 3,
-                    "unit": "steps",
-                    "progress": None,
-                    "desc": "alphabet",
-                }
-            ],
-        ]
-
-
 class TestAddRequests:
     def test_no_type_hints(self):
         def moo(a, b):
